+++ CHANGELOG.md
@@ -8,7 +8,9 @@ Go to the `v1` branch to see the changelog of Lume 1.
 ## [2.0.3] - Unreleased
 ### Fixed
 - `sitemap` plugin: Add the `xmlns` namespace for localized urls.
-- `multilanguage` plugin: Fix error of two pages with the same id, type and lang.
+- `multilanguage` plugin:
+  - Fix error of two pages with the same id, type and lang.
+  - Fix the error of a page with lang, but undefined id.
 
 ## [2.0.1] - 2023-12-10
 ### Added