+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Go to the `v1` branch to see the changelog of Lume 1.
   - Fix error of two pages with the same id, type and lang.
   - Fix the error of a page with lang, but undefined id.
 - Removed unused `imagick` dependency.
+- Added `Lume.PaginateResult` type.
 
 ## [2.0.1] - 2023-12-10
 ### Added